Add Molecule representation and update package deployment (#18)
* Update visualization data file

* Update github actions workflows

* Add Molecule class to represent molecular structure with different surface representation and calculate molecular volume

* Update publish workflow

* Update requirements.txt

* Add Molecule to documentation

* Add Molecule to documentation and refactor text

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Add pyproject.toml to pyKVFinder package

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update test files and README

* Update test files and README

* Update test files and README

* Update test files and README

* Update test files and README

* Update test files and README

* Update test files and README

* Update workflows

* Update workflows

* Update workflows

* Update workflows

* Update testing workflows

* Add coverage workflow

* Update coverage workflow

* Update coverage workflow

* Update README

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update testing workflow

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Update unit testing

* Add optional dependencies

* Update unit testing

* Add tests for get_vertices_from_file

* Add tests for get_vertices_from_file

* Add tests for detect

* Add tests for detect

* Add tests for spatial

* Add tests for depth

* Add tests for constitutional

* Add tests for constitutional

* Add tests for hydropathy

* Update unit tests

* Add tests for openings and templates for export, process_openings, openings, export_openings

* Remove tests for prints of verbose

* Add test for export

* Update unit testing

* Update unit testing

* Update integration testing

* Update tests

* Update publish release workflow

* Update publish release workflow

* Update publish release workflow

* Update publish-release.yml

* Update publish-release.yml

* Update test_grid.py

* Update pyproject.toml

* New build and deployment workflow

* New build and deployment workflow

* New build and deployment workflow

* New build and deployment workflow

* New build and deployment workflow

* New build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update build and deployment workflow

* Update area calculation routine

* Update deploy.yml

* Fix voxel classification in area estimation

* Update version

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update installation procedure

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflow

* Update workflows

* Update package dependencies

* Update testing

* Update testing

+++ C/pyKVFinder.h
@@ -94,7 +94,9 @@ double get_hydrophobicity_value(char *resname, char **resn, double *scales, int 
 void project_hydropathy(double *hydropathy, int *surface, int nxx, int nyy, int nzz, double *atoms, int natoms, int xyzr, double *reference, int ndims, double *sincos, int nvalues, char **resname, char **resn, double *scales, int nscales, double step, double probe_in, int nthreads);
 
 /* Estimate average hydropathy */
-void estimate_average_hydropathy(double *avgh, int ncav, double *hydropathy, double *surface, int nx, int ny, int nz, int nthreads);
+void estimate_average_hydropathy(double *avgh, int ncav, double *hydropathy,
+                                 int *surface, int nx, int ny, int nz,
+                                 int nthreads);
 
 /* Hydropathy characterization */
 void _hydropathy(double *hydropathy, int size, double *avgh, int ncav, int *surface, int nxx, int nyy, int nzz, double *atoms, int natoms, int xyzr, double *reference, int ndims, double *sincos, int nvalues, char **resname, char **resn, double *scales, int nscales, double step, double probe_in, int nthreads, int verbose);
